Reference

SpmGrids.get_channel

SpmGrids.get_parameter

SpmGrids.load_grid

SpmGrids.plot_line

SpmGrids.plot_spectrum

SpmGrids.get_channel — Function

get_channel(grid::SpmGrid, name::AbstractString,
    index_x::GridRange, index_y::GridRange, index_channel::GridRange=:)::Array{Float32}

Returns the data for the channel name at the point(s) specified by index_x, index_y The channel data can be indexed by index_channel.

SpmGrids.get_parameter — Method

get_parameter(grid::SpmGrid, name::AbstractString,
    index_x::GridRange, index_y::GridRange)::Union{Float32, Array{Float32}}

Returns the value for parameter name at the point(s)specified by index_x, index_y.

SpmGrids.load_grid — Method

load_grid(filename::AbstractString; header_only::Bool=false)

Loads a grid from a binary file. If header_only is true, then only the header is loaded.

SpmGrids.plot_line — Function

plot_line(grid::SpmGrid, channel_y::String,
    index_x::GridRange, index_y::GridRange, index_channel::GridRange=nothing;
    channel_x::String="", backward::Bool=true, ax::Any=nothing, backend::Module=Main,
    kwargs...)::Nothing)

Plots a line in the three-dimensional data spanned by xy plane and the spectroscopy data channel_y. Indexing is done through index_x, index_y and index_channel and should be done such that a one-dimensional array is obtained. It is also possible to plot channel_y vs channel_x (which defaults ot the sweep signal if not specified) for one point in the grid If include_backward is true` (default), the plot will include data from backward sweep as well (if they exist).

Before using this function, a Makie backend (GLMakie, CairoMakie or WGLMakie) should be imported and the figure or axis should be set up. A particular Axis can be specified via the ax keyword argument. By default, the Makie backend from the Main module is used; it can also be directly specified via the backend keyword argument.

Extra keyword arguments can be specified and will be passed through to the plot function. Keyword arrguments with the suffix _bwd will be used for plotting of the backward scan.

SpmGrids.plot_spectrum — Function

plot_spectrum(grid::SpmGrid, channel_x::String, channel_y::String,
    index_x::GridRange, index_y::GridRange, index_channel::GridRange=:;
    backward::Bool=true, ax::Any=nothing, backend::Module=Main,
    kwargs...)::Nothing)

Plots a scatter plot of channel_y vs channel_x on the given index_x and y_index. If channel_x is "", then the sweep signal will be used for channel_x. Additionally, the spectrum data can be indexed by index_channel. If include_backward is true` (default), the plot will include data from backward sweep as well (if they exist).

Before using this function, a Makie backend (GLMakie, CairoMakie or WGLMakie) should be imported and the figure or axis should be set up. A particular Axis can be specified via the ax keyword argument. By default, the Makie backend from the Main module is used; it can also be directly specified via the backend keyword argument.

Extra keyword arguments can be specified and will be passed through to the plot function. Keyword arrguments with the suffix _bwd will be used for plotting of the backward scan.