Reference

SpmGrids.get_channel

SpmGrids.get_parameter

SpmGrids.load_grid

SpmGrids.get_channel — Function

get_channel(grid::SpmGrid, name::AbstractString, index_x::Any, index_y::Any, index_channel::Any=nothing)::Array{Float32}

Returns the data for the channel name at the point(s) specified by index_x, index_y The channel data can be indexed by index_channel.

SpmGrids.get_parameter — Method

get_parameter(grid::SpmGrid, name::AbstractString, index_x::Any, index_y::Any)::Union{Float32, Array{Float32}}

Returns the value for parameter name at the point(s)specified by index_x, index_y.

SpmGrids.load_grid — Method

load_grid(filename::AbstractString; header_only::Bool=false)

Loads a grid from a binary file. If header_only is true, then only the header is loaded.