Reference

SpmGrids.get_channel

SpmGrids.get_parameter

SpmGrids.load_grid

SpmGrids.plot_cube

SpmGrids.plot_line

SpmGrids.plot_plane

SpmGrids.plot_spectrum

SpmGrids.xyindex_to_point

SpmGrids.get_channel — Function

get_channel(grid::SpmGrid, name::AbstractString,
    x_index::GridRange, y_index::GridRange, channel_index::GridRange=:)::Array{Float32}

Returns the data for the channel name at the point(s) specified by x_index, y_index The channel data can be indexed by channel_index.

SpmGrids.get_parameter — Method

get_parameter(grid::SpmGrid, name::AbstractString,
    x_index::GridRange, y_index::GridRange)::Union{Float32, Array{Float32}}

Returns the value for parameter name at the point(s)specified by x_index, y_index.

SpmGrids.load_grid — Method

load_grid(filename::AbstractString; header_only::Bool=false)

Loads a grid from a binary file. If header_only is true, then only the header is loaded.

SpmGrids.plot_cube — Function

function plot_cube(grid::SpmGrid, response_channel::String,
    x_index::GridRange, y_index::GridRange, channel_index::GridRange=nothing;
    backward::Bool=false, ax::Any=nothing, backend::Module=Main,
    kwargs...)::Tuple{Any,String}

Plots a cube of response_channel in the three-dimensional data spanned by the x,y plane and the sweep signal. Indexing is done through x_index, y_index and channel_index and should be done such that a three-dimensional array is obtained. If backward is set to true, then data from the backward sweep is plotted if it exists.

Before using this function, a Makie backend (GLMakie, CairoMakie or WGLMakie) should be imported and the figure or axis should be set up. A particular Axis can be specified via the ax keyword argument. By default, the Makie backend from the Main module is used; it can also be directly specified via the backend keyword argument.

Extra keyword arguments can be specified and will be passed through to the plot function.

Returns a tuple containing the volume-plot, and a colorbar label.

SpmGrids.plot_line — Function

plot_line(grid::SpmGrid, response_channel::String,
    x_index::GridRange, y_index::GridRange, channel_index::GridRange=nothing;
    sweep_channel::String="", backward::Bool=true, ax::Any=nothing, backend::Module=Main,
    kwargs...)::Nothing

Plots the response_channel along a line in the three-dimensional data spanned by x,y plane and the spectroscopy data. Indexing is done through x_index, y_index and channel_index and should be done such that a one-dimensional array is obtained. It is also possible to plot response_channel vs sweep_channel (which defaults to the sweep signal if not specified) for one point in the grid If backward is true (default), the plot will include data from backward sweep as well (if they exist).

Before using this function, a Makie backend (GLMakie, CairoMakie or WGLMakie) should be imported and the figure or axis should be set up. A particular Axis can be specified via the ax keyword argument. By default, the Makie backend from the Main module is used; it can also be directly specified via the backend keyword argument.

Extra keyword arguments can be specified and will be passed through to the plot function. Keyword arrguments with the suffix _bwd will be used for plotting of the backward scan.

SpmGrids.plot_plane — Function

plot_plane(grid::SpmGrid, response_channel::String,
    x_index::GridRange, y_index::GridRange, channel_index::GridRange=nothing;
    backward::Bool=false, ax::Any=nothing, backend::Module=Main,
    kwargs...)::Tuple{Any, String}

Plots a plane of response_channel in the three-dimensional data spanned by x,y plane and the seep signal. Indexing is done through x_index, y_index and channel_index and should be done such that a two-dimensional array is obtained. If backward is set to true, then data from the backward sweep is plotted if it exists.

Before using this function, a Makie backend (GLMakie, CairoMakie or WGLMakie) should be imported and the figure or axis should be set up. A particular Axis can be specified via the ax keyword argument. By default, the Makie backend from the Main module is used; it can also be directly specified via the backend keyword argument.

Extra keyword arguments can be specified and will be passed through to the plot function.

Returns a tuple containing the heatmap, the colorbar label and the plot label.

SpmGrids.plot_spectrum — Function

plot_spectrum(grid::SpmGrid, sweep_channel::String, response_channel::String,
    x_index::GridRange, y_index::GridRange, channel_index::GridRange=:;
    backward::Bool=true, ax::Any=nothing, backend::Module=Main,
    kwargs...)::Nothing

Plots a line plot of response_channel vs sweep_channel on the given x_index and y_index. If sweep_channel is "", then the sweep signal will be used for sweep_channel. Additionally, the spectrum data can be indexed by channel_index. If include_backward is true` (default), the plot will include data from backward sweep as well (if they exist).

Before using this function, a Makie backend (GLMakie, CairoMakie or WGLMakie) should be imported and the figure or axis should be set up. A particular Axis can be specified via the ax keyword argument. By default, the Makie backend from the Main module is used; it can also be directly specified via the backend keyword argument.

Extra keyword arguments can be specified and will be passed through to the plot function. Keyword arrguments with the suffix _bwd will be used for plotting of the backward scan.

SpmGrids.xyindex_to_point — Method

xyindex_to_point(grid::SpmGrid, index_x::Int, index_y)::Tuple{Float32,Float32}

Converts index_x and index_y of grid to point coordinates in physical units.